# apps/manager/docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    email: support@sysarmor.com
    name: SysArmor Team
    url: https://sysarmor.com/support
  description: SysArmor EDR 系统的控制平面服务 API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://sysarmor.com/terms
  title: SysArmor Manager API
  version: "1.0"
paths:
...
  /services/opensearch/cluster/stats:
    get:
      consumes:
      - application/json
      description: 获取 OpenSearch 集群的详细统计信息，包括索引、节点、存储等统计
      produces:
      - application/json
      responses:
        "200":
          description: 集群统计信息
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取 OpenSearch 集群统计信息
      tags:
      - opensearch
  /services/opensearch/events/aggregations:
    get:
      consumes:
      - application/json
      description: 获取安全事件的聚合统计信息，包括按数据源、严重程度、事件类型等维度的统计
      parameters:
      - description: 索引模式，默认为 sysarmor-events-*
        in: query
        name: index
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 聚合统计结果
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取事件聚合统计
      tags:
      - opensearch
  /services/opensearch/events/by-source:
    get:
      consumes:
      - application/json
      description: 获取来自指定数据源的安全事件
      parameters:
      - description: 索引模式，默认为 sysarmor-events-*
        in: query
        name: index
        type: string
      - description: 数据源名称
        in: query
        name: source
        required: true
        type: string
      - description: 返回结果数量，默认为10，最大100
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 事件列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 根据数据源获取安全事件
      tags:
      - opensearch
  /services/opensearch/events/high-risk:
    get:
      consumes:
      - application/json
      description: 获取风险评分高于指定阈值的安全事件
      parameters:
      - description: 索引模式，默认为 sysarmor-events-*
        in: query
        name: index
        type: string
      - description: 最小风险评分
        in: query
        name: min_score
        required: true
        type: integer
      - description: 返回结果数量，默认为10，最大100
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 高风险事件列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 根据风险评分获取安全事件
      tags:
      - opensearch
  /services/opensearch/events/recent:
    get:
      consumes:
      - application/json
      description: 获取最近一段时间内的安全事件
      parameters:
      - description: 索引模式，默认为 sysarmor-events-*
        in: query
        name: index
        type: string
      - description: 时间范围（小时），默认为24小时
        in: query
        name: hours
        type: integer
      - description: 返回结果数量，默认为10，最大100
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 最近事件列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取最近的安全事件
      tags:
      - opensearch
  /services/opensearch/events/search:
    get:
      consumes:
      - application/json
      description: 在指定索引模式中搜索安全事件
      parameters:
      - description: 索引模式，默认为 sysarmor-events-*
        in: query
        name: index
        type: string
      - description: 搜索查询字符串
        in: query
        name: q
        type: string
      - description: 返回结果数量，默认为10，最大100
        in: query
        name: size
        type: integer
      - description: 结果偏移量，默认为0
        in: query
        name: from
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 搜索结果
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 搜索安全事件
      tags:
      - opensearch
  /services/opensearch/events/threats:
    get:
      consumes:
      - application/json
      description: 获取标记为可疑的威胁事件
      parameters:
      - description: 索引模式，默认为 sysarmor-events-*
        in: query
        name: index
        type: string
      - description: 返回结果数量，默认为10，最大100
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 威胁事件列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取威胁事件
      tags:
      - opensearch
  /services/opensearch/events/time-range:
    get:
      consumes:
      - application/json
      description: 获取指定时间范围内的安全事件
      parameters:
      - description: 索引模式，默认为 sysarmor-events-*
        in: query
        name: index
        type: string
      - description: 开始时间 (RFC3339 格式)
        in: query
        name: from
        required: true
        type: string
      - description: 结束时间 (RFC3339 格式)
        in: query
        name: to
        required: true
        type: string
      - description: 返回结果数量，默认为10，最大100
        in: query
        name: size
        type: integer
      - description: 页码，默认为1
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 事件列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 根据时间范围获取安全事件
      tags:
      - opensearch
  /services/opensearch/indices:
    get:
      consumes:
      - application/json
      description: 获取 OpenSearch 集群中所有索引的信息
      produces:
      - application/json
      responses:
        "200":
          description: 索引列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取 OpenSearch 索引列表
      tags:
      - opensearch
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"